va_platform-master/services/getfindings/requirements.txt
flask>=1.1.1
git+https://github.com/gva-jjoyce/gva_data_validator
git+https://github.com/gva-jjoyce/gva_logging
git+https://github.com/gva-jjoyce/gva_data_flows
git+https://github.com/joocer/gva_services
google-cloud-storage
